Bug Fixes (Move Detection and Shuffle)
The game no longer incorrectly displays "out of moves". (For a deck of
size four with two suits in the pattern ABBA, the program displayed "out
of moves" and didn't consider the obvious winning move.) The RNG is now
more fair. (Before when shuffling the deck, the last card would never
get shuffled.)